requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
